IMPORTANT: Breaking change.
Message{} and Version{} interface methods are changed.

The rationale behind this is that, sometimes we may not know the size of
encoded Message apriori. `fixbuffer()` is a convenient logic to manage
a buffer that can automatically adjust to maximum size.

Had to fix a whole bunch of files to deal with this. But both unit-tests
and benchmarks are passing fine.